+++ plugins/inputs/container/k8s.go
@@ -7,8 +7,6 @@ package container
 
 import (
 	"fmt"
-	"net/url"
-	"strings"
 
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/io/point"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs"
@@ -50,7 +48,7 @@ func (k *kubernetesInput) gatherResourceMetric() (inputsMeas, error) {
 	)
 
 	for _, fn := range k8sResourceMetricList {
-		x := fn(k.client, k.ipt.Tags, k.getHost())
+		x := fn(k.client, k.ipt.Tags)
 
 		if xPod, ok := x.(podResourceInterface); ok {
 			xPod.setExtractK8sLabelAsTags(k.ipt.ExtractK8sLabelAsTags)
@@ -106,7 +104,7 @@ func (k *kubernetesInput) gatherResourceObject() (inputsMeas, error) {
 	)
 
 	for _, fn := range k8sResourceObjectList {
-		x := fn(k.client, k.ipt.Tags, k.getHost())
+		x := fn(k.client, k.ipt.Tags)
 		if m, err := x.object(k.ipt.Election); err == nil {
 			res = append(res, m...)
 		} else {
@@ -121,24 +119,10 @@ func (k *kubernetesInput) watchingEventLog(done <-chan interface{}) {
 	watchingEvent(k.client, k.ipt.Tags, done, k.ipt.Election)
 }
 
-func (k *kubernetesInput) getHost() string {
-	u := k.ipt.K8sURL
-	if strings.Contains(u, "127.0.0.1") || strings.Contains(u, "localhost") {
-		return ""
-	}
-	uu, err := url.Parse(u)
-	if err != nil {
-		l.Errorf("getHost: failed to parse k8s URL: %v", err)
-		return ""
-	}
-	return uu.Host
-}
-
 type k8sResourceMetricInterface interface {
 	name() string
 	metric(election bool) (inputsMeas, error)
 	count() (map[string]int, error)
-	getHost() string
 }
 
 type k8sResourceObjectInterface interface {
@@ -186,8 +170,8 @@ func defaultNamespace(ns string) string {
 }
 
 type (
-	newK8sResourceMetricHandle func(k8sClientX, map[string]string, string) k8sResourceMetricInterface
-	newK8sResourceObjectHandle func(k8sClientX, map[string]string, string) k8sResourceObjectInterface
+	newK8sResourceMetricHandle func(k8sClientX, map[string]string) k8sResourceMetricInterface
+	newK8sResourceObjectHandle func(k8sClientX, map[string]string) k8sResourceObjectInterface
 )
 
 var (